Add keyboard shortcut for zsh push-input
This is based on the function found at [1]. Now Ctrl + Z can be used to
interrupt typing a command and do something else. After the other
command completes, the prompt will be brought up again with the
unfinished prompt content of before pressing Ctrl + Z, an awesome
feature of zsh.

[1]: https://unix.stackexchange.com/questions/10825/remember-a-half-typed-command-while-i-check-something/10851#10851