Add git-whatsnew command
When run after a pull, this will list the --oneline message of every new
commit. This does include merge commits. For the Chromium repository
that is not an issue because it has a (mostly) linear history, but for
e.g. Rust and Servo I might want to exclude merge commits in the future,
as about half of the commits are merge commits in those repositories.