List some more Arch packages
Some of these are dependencies of others, but I would like to keep them
in the list anyway because I use them explicitly. These packages should
also be installed explicitly, not as dependencies. Packages not in the
list *should* be installed as dependencies.

The list was generated using

    $ pacman -Qneq > packages

I have a few more packages installed which are specific to this machine,
so I probably should not list them here, or at least I should split them
off into a different list.