Enable relative line numbers in Vim
Learned from a blog post by Petr Zemek. I wonder how this pans out, the
line numbers are a lot of visual clutter, but perhaps it is worth the
boost in navigation speed. Especially now that I have enabled hardtime,
which prevents me from repeatedly mashing on jk. Often the best
solution I know of is to count beforehand how many lines I need to move,
and relative line numbers make it easy to do that in a glance.

They do look ugly though, especially on the command line.