Update elm.vim plugin to fix text width breakage
For a while, text wrapping when editing Git commit messages was broken.
Upon further investigation, the last thing I did in my Vim config was to
install the elm.vim plugin. It turns out this was a known problem (or at
least, a different problem was known) upstream, and it has been fixed.
After updating line wrapping is now enabled again by default.