Make gVim window borders less ugly in GTK3
This only works for the dark window thoug, when I :set bg=light this
breaks. But it is better than the gray borders I guess. What I really
want is for GTK to just snap the window size to whatever Vim wants, and
not allow the extra border in the first place.