Add Python script that tree-tabulates "git branch"
Because I want proper table alignment. It can be done with a --format
argument to "git branch", but that requires hard-coding column widths.
This script computes the minimal width per column and uses that, while
still aligning.

Furthermore, this script orders branches behind their upstream if they
have one, indented by two spaces, to clarify the tree structure of the
branches. Note that this is upstreams, not ancestors! I hope this will
allow me to keep my sanity while working on a dozen dependent feature
branches, keeping track of what to rebase.