Fail to derive monoid in git-br.hs
This was the thing that made me expect that I could write this nicely in
Haskell, but it turns out not to be possible, unfortunately.

I want "Branch a" to be a monoid if "a" is, then I could first map every
branch to "Branch [a]" by putting fields in singleton lists, and then I
could foldMap them all together to get the columns. But alas, records in
Haskell are not named tuples, this does not work.