Add alias for setting upstream to master
I use this often, when I fork a branch off some other branch, and I
default to setting the upstream to that branch, but then later I merge
the parent branch, and I want to set the upstream of the child to master
and rebase it onto master. This alias is useful for that.