Begin working on btrfs defrag script
My awk script in my notes lists both directories and files, causing some
things to be defragged twice. Rather than trying to save that thing,
rewrite it in Python already.

Interestingly, running "btrfs filesystem du" is slow, but defrag is slow
too. We could pipeline the two with a channel. But in Python this is not
easy to do, as far as I am aware. Perhaps I'll rewrite this in Haskell
some time.