Begin working on script to sync btrfs subvolumes
I have two btrfs filesystems that contain snapshots named after dates.
I want to sync the two in a way that preserves as much sharing as
possible. But I don't want to use btrfs send/receive, first of all
because I don't trust it, and second, because I want to keep the two
file systems as isolated as possible, such that a bug in one doesn't
get synced to the second one, corrupting both.