Add Ale and Rust config for language server tools
The autocomplete doesn't seem to work very well, and the linter
squiggles are annoying at best. The hover tooltip looks ugly (though
that's more gVim's fault) and doesn't always work.

But, the sole thing that is going to instantly make up for all of this
is ... working jump to definition!