Add CUE splitting script
I used to use "shnsplit" for this, until I ran into some files that it
couldn't handle. Then I found some Bash multi-liners on the interwebs
based on "cuebreakpoints" and "cueprint". It worked, but it is cryptic,
Bash is not not a great language for things that need to be longer than
3 lines. So rewrite it in Python instead.