Pin Nix to 2.3
Nix 2.4 broke the entire CLI, and although it was always marked as
"experimental", it had been stable for years, and it worked great. In
Nix 2.4 suddenly none of the "nix run" commands work any more, and there
doesn't appear to be an easy alternative. I don't want to deal with all
of this right now, I just want to get my development environment
working, so I'm going to pin things to Nix 2.3.