Do not make GC root for flake source itself
I see Nix rebuilding my devenvs all the time, and I think it happens
because now I take all the flake sources to turn into GC roots, but
the flake sources include the source code of the repository itself,
which changes on every commit. So exclude that one, which should be
easy enough because it is top-level but the other ones are nested in
"inputs".