Make zsh record history timestamps
As I learned from [1], it seems like a nice improvement to have relative
time offsets in my history. As a first step, save timestamps. While I'm
at it, bump the history size further. I did bump it once in the past
because it was really too low, now it's basically fine, but also making
it larger doesn't really hurt so let's make it larger.

[1]: https://tratt.net/laurie/blog/2025/better_shell_history_search.html