Begin working on custom Arch mirror ranker
I've been using Reflector, but every time I reconfigure my mirrorlist,
it's fine, and then a week later it slows to a crawl, and I run
Reflector again, but I never saturate my uplink. It turns out, I think
Reflector does not _measure_ which mirror is the fastest, it just trusts
the Arch mirror checker, but then we get their point of view, which may
be very different from mine.

So let's build a little tool that actually measures which mirrors are
the fastest.