Helix: Add shorthand for sorting selected lines
I do this often. It's easy to do in Vim and I find it annoying in Helix,
because you have to split the selection first, and that uses Alt which
is annoying to type on my keyboard. But we can just configure a
shortcut! It can do the splitting and sorting at the same time!