Rewrite display brightness script in Python
I had this script for setting brightness with ddcutil for some time.
It was terribly slow, and for some reason I just accepted that. Then I
tried KDE Plasma, and it has brightness sliders that respond instantly.
So it's not *inherently* slow. Thanks also to David for noting that
ddcutil can be made faster.

So why is ddcutil so slow? Because when you pass --display, it has to
first detect all displays so it can give them a number. As the FAQ at
<https://www.ddcutil.com/faq/> says:

> Typically, ddcutil spends 90% of its elapsed time in waits mandated
> by the DDC/CI protocol. (..) If the /dev/i2c bus number of the monitor
> is given, ddcutil skips the initial phase of searching the /dev/i2c
> devices for monitors.

The /dev/i2c-* numbers are not stable across reboots, or even turning
displays on and off, but it is still possible (at least with my setup)
to find which ones are related to a display by just poking around a bit
in the /dev and /sys filesystems, which is very fast.

So now I have a script that responds ~instantly, instead of waiting for
2 seconds, setting brightness of one display, waiting 2 seconds again,
and doing the second one. And it doesn't even require root!