Add expbackup script
I have been using this for more than a year now to determine which btrfs
snapshots to delete when I snapshot my root partition, but it applies
more widely. I still need to write a blog post about it, but in the
meantime, it's useful to have in other places. It's just a few lines,
I'll put it in here, not worth a separate repo.